changed the layout of the application
