+++ Frontend/src/components/icons/RedditIcon.vue
@@ -1,9 +1,9 @@
 <template>
-    <div class="ml-5 py-2 cursor-pointer">
-        <svg version="1.1" class="w-28" id="Layer_1" xmlns="http://www.w3.org/2000/svg"
-             xmlns:xlink="http://www.w3.org/1999/xlink"
-             x="0px" y="0px"
-             viewBox="0 0 2500 810" style="enable-background:new 0 0 2500 810;" xml:space="preserve">
+
+    <svg version="1.1" class="w-28" id="Layer_1" xmlns="http://www.w3.org/2000/svg"
+         xmlns:xlink="http://www.w3.org/1999/xlink"
+         x="0px" y="0px"
+         viewBox="0 0 2500 810" style="enable-background:new 0 0 2500 810;" xml:space="preserve">
 
 <g>
 	<circle class="st0" fill="#FF4500" cx="406.6" cy="405.6" r="402.3"/>
@@ -18,33 +18,33 @@
 		c4.4-4.3,11.6-4.2,15.9,0.2s4.2,11.6-0.2,15.9l0,0L506.3,556.5L506.3,556.5z M499.1,487.6c-22.2,0-40.3-18.1-40.3-40.3
 		c0-22.2,18.1-40.3,40.3-40.3c22.2,0,40.3,18.1,40.3,40.3c0.9,22.2-16.4,40.9-38.6,41.8c-0.7,0-1.3,0-2,0L499.1,487.6z"/>
 </g>
-            <g>
+        <g>
 	<circle class="st0" fill="white" cx="2248.8" cy="228.1" r="54.8"/>
-                <path fill="white" d="M1472.2,461.3c21.2,1.5,39.6-14.5,41.2-35.8c0.1-1.2,0.1-2.6,0.1-3.9c-0.3-7.4-1.2-14.9-2.7-22.2
+            <path fill="white" d="M1472.2,461.3c21.2,1.5,39.6-14.5,41.2-35.8c0.1-1.2,0.1-2.6,0.1-3.9c-0.3-7.4-1.2-14.9-2.7-22.2
 		c-13-69.4-72.3-120.5-142.8-123.3c-80.3,0-147.4,73.8-147.4,164.4s66,164.5,147.3,164.5c43.1,1.7,84.9-15.7,113.9-47.6
 		c11.9-13,10.9-33.1-2.1-45c-1-0.9-1.9-1.7-2.9-2.4c-13-8.1-29.9-6.4-40.9,4.2c-18.1,18-42.4,28.3-68,29.1
 		c-44.2-3.5-79.6-38.1-84-82.4H1471L1472.2,461.3z M1367.6,338c38.9,2.6,71.6,30.4,80.3,68.5h-160.4
 		C1296.1,368.3,1328.6,340.5,1367.6,338z"/>
-                <path fill="white" d="M1223.8,308c0-15.4-11.4-28.5-26.7-30.4c-41.3-6.9-83.3,8.4-110.4,40.4v-4c0-17.1-13.8-30.9-30.9-30.9
+            <path fill="white" d="M1223.8,308c0-15.4-11.4-28.5-26.7-30.4c-41.3-6.9-83.3,8.4-110.4,40.4v-4c0-17.1-13.8-30.9-30.9-30.9
 		c-17.1,0-30.9,13.8-30.9,30.9v253c-0.4,16.6,12.3,30.7,28.9,32.1c16.9,1.2,31.7-11.6,32.8-28.5c0.1-1.2,0.1-2.4,0-3.6V437.2
 		c-2.7-51.8,37-95.9,88.8-98.7c4.6-0.3,9.3-0.2,13.9,0.3h6.2C1211.5,337.4,1223.8,324.1,1223.8,308z"/>
-                <path fill="white" d="M2279.6,341.2c0-17.1-13.8-30.9-30.9-30.9c-17.1,0-30.9,13.8-30.9,30.9l0,0v225.7c0,17.1,13.8,30.9,30.9,30.9
+            <path fill="white" d="M2279.6,341.2c0-17.1-13.8-30.9-30.9-30.9c-17.1,0-30.9,13.8-30.9,30.9l0,0v225.7c0,17.1,13.8,30.9,30.9,30.9
 		c17.1,0,30.9-13.8,30.9-30.9V341.2z"/>
-                <path fill="white" d="M1823.6,169.7c0-17-13.8-30.8-30.8-30.8l0,0l0,0c-17,0-30.8,13.8-30.8,30.8v138c-20.8-21-49.4-32.4-78.9-31.6
+            <path fill="white" d="M1823.6,169.7c0-17-13.8-30.8-30.8-30.8l0,0l0,0c-17,0-30.8,13.8-30.8,30.8v138c-20.8-21-49.4-32.4-78.9-31.6
 		c-80.3,0-147.4,73.8-147.4,164.4s66.1,164.5,147.4,164.5c30,0.9,59-10.6,80.3-31.9c4.8,16.3,21.9,25.6,38.2,20.8
 		c12.3-3.6,21.1-14.4,21.9-27.3L1823.6,169.7L1823.6,169.7z M1682.9,544.1c-47.3,0-85.8-46-85.8-102.9s38.2-102.9,85.8-102.9
 		c47.6,0,85.8,46,85.8,102.9s-38.2,102.4-85.6,102.4L1682.9,544.1z"/>
-                <path fill="white" d="M2145.8,169.7c0-17-13.8-30.8-30.8-30.8l0,0c-17,0-30.8,13.8-30.8,30.8l0,0v138c-20.8-21-49.4-32.4-78.9-31.6
+            <path fill="white" d="M2145.8,169.7c0-17-13.8-30.8-30.8-30.8l0,0c-17,0-30.8,13.8-30.8,30.8l0,0v138c-20.8-21-49.4-32.4-78.9-31.6
 		c-80.3,0-147.4,73.8-147.4,164.4s66.1,164.5,147.4,164.5c30,0.9,59-10.6,80.3-31.9c4.8,16.3,21.9,25.6,38.2,20.8
 		c12.3-3.6,21.1-14.4,21.9-27.3L2145.8,169.7L2145.8,169.7z M2005.2,544.1c-47.3,0-85.8-46-85.8-102.9s38.2-102.9,85.8-102.9
 		c47.6,0,85.8,46,85.8,102.9s-38.2,102.4-85.6,102.4L2005.2,544.1z"/>
-                <path fill="white" d="M2439.7,566.8V338h26.7c14.8,0.9,27.4-10.4,28.3-25.1c0-0.5,0.1-1.1,0.1-1.7c1.3-14.7-9.6-27.7-24.2-29
+            <path fill="white" d="M2439.7,566.8V338h26.7c14.8,0.9,27.4-10.4,28.3-25.1c0-0.5,0.1-1.1,0.1-1.7c1.3-14.7-9.6-27.7-24.2-29
 		c-0.8-0.1-1.7-0.1-2.5-0.1h-28.3v-42.5c0.2-16.5-12.4-30.4-28.9-31.9c-16.9-1.2-31.6,11.6-32.8,28.5c-0.1,0.7-0.1,1.5-0.1,2.2v44.4
 		h-26.8c-14.8-0.9-27.4,10.4-28.3,25.1c0,0.5-0.1,1.1-0.1,1.7c-1.3,14.7,9.6,27.7,24.2,29c0.8,0.1,1.7,0.1,2.5,0.1h26.7v228.2
 		c0,17,13.8,30.8,30.8,30.8l0,0l0,0c17,1.1,31.5-11.9,32.6-28.9C2439.7,568.2,2439.7,567.5,2439.7,566.8z"/>
 </g>
 </svg>
-    </div>
+
 </template>
 <script lang="ts">
 export default {
