+++ Frontend/src/router/index.ts
@@ -1,14 +1,18 @@
 import { createRouter, createWebHistory } from 'vue-router'
-import HomeView from "@/views/HomeView.vue";
+import HomeView from "@/views/MainLayout.vue";
 
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
   routes: [
     {
       path: '/',
-      name: 'home',
       component: HomeView,
       children: [
+        {
+          path: '',
+          name: 'home',
+          component: () => import('../views/HomeView.vue')
+        },
         {
           path: '/r/:community/comments/:id',
           name: 'comments',
@@ -24,7 +28,7 @@ const router = createRouter({
     {
       path: '/test',
       name: 'test',
-      component: () => import('../views/HomeView.vue')
+      component: () => import('../views/MainLayout.vue')
     },
 
   ]
